UdpTest.jl

UdpTest.jl

Documentation for UdpTest.jl

An Example for Module porting from Python to Julia language

UdpTest.client — Method.

function client(host::String, port::Int)
require a message to send to the HOST on the PORT.
    Eventually show the received replay.

source

UdpTest.server — Method.

function server(port::Int) start to listen on port and reply any Received message

source